fix(windows-remote-desktop): compare keyboard flags as DWORD
The new unicode-text test compared KEYINPUT.dwFlags (unsigned long) against
the KEYEVENTF_* macros, which expand to int. Chromium's clang build turns
-Wsign-compare into an error, so the CI windows worker job failed to compile
input_injector_unittest.cc while the local MSVC run accepted it.

Cast the expected flags to DWORD and include <string> for the desktop-name
helper.

Reproduced the exact CI error with the Chromium clang toolchain, then
confirmed the fix: input_injector_unittests compiles and runs 9 passed,
1 skipped (the Winlogon-only input probe), and display_capture_unittests,
worker_policy_unittests and the worker binary still build clean.

+++ native/windows-remote-desktop/input_injector_unittest.cc
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <cwchar>
 #include <iterator>
+#include <string>
 #include <vector>
 
 #include "test/gtest.h"
@@ -191,10 +192,11 @@ TEST(InputArbiterTest, EncodesTextAsBoundedUnicodeInputPairs) {
   EXPECT_TRUE(input.Text(u"A"));
   ASSERT_EQ(recording.events.size(), 2u);
   EXPECT_EQ(recording.events[0].ki.wScan, static_cast<WORD>(u'A'));
-  EXPECT_EQ(recording.events[0].ki.dwFlags, KEYEVENTF_UNICODE);
+  EXPECT_EQ(recording.events[0].ki.dwFlags,
+            static_cast<DWORD>(KEYEVENTF_UNICODE));
   EXPECT_EQ(recording.events[1].ki.wScan, static_cast<WORD>(u'A'));
   EXPECT_EQ(recording.events[1].ki.dwFlags,
-            KEYEVENTF_UNICODE | KEYEVENTF_KEYUP);
+            static_cast<DWORD>(KEYEVENTF_UNICODE | KEYEVENTF_KEYUP));
 }
 
 TEST(InputArbiterTest, SecureConsoleAcceptsNonTextInputWithoutChangingState) {
